requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
docstring-parser==0.16
executing==2.1.0
fastapi==0.114.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
icecream==2.1.3
idna==3.8
instructor==1.4.1
jiter==0.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
openai==1.44.1
pillow==10.4.0
pip @ file:///home/conda/feedstock_root/build_artifacts/pip_1724954388421/work
pydantic==2.9.1
pydantic-core==2.23.3
pygments==2.18.0
requests==2.32.3
rich==13.8.1
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
tenacity==8.5.0
tqdm==4.66.5
typer==0.12.5
typing-extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
yarl==1.11.1
